eCommerce Yubit - Product Composite Service - Calling other services

+++ api/src/main/java/com/yubit/api/core/composite/product/ProductAggregate.java
@@ -3,7 +3,7 @@ package com.yubit.api.core.composite.product;
 import java.util.List;
 
 public record ProductAggregate(
-        Long productId,
+        int productId,
         String name,
         int weight,
         List<RecommendationSummary> recommendations,
