+++ api/src/main/java/com/yubit/api/core/composite/product/ProductCompositeService.java
@@ -11,5 +11,5 @@ public interface ProductCompositeService {
             value = "/product-composite/{productId}",
             produces = "application/json"
     )
-    ProductAggregate getProduct(@PathVariable Long productId);
+    ProductAggregate getProduct(@PathVariable int productId);
 }
