+++ api/src/main/java/com/yubit/api/core/composite/product/ReviewSummary.java
@@ -4,7 +4,7 @@ package com.yubit.api.core.composite.product;
  * especificamos lo que nos interesa del Review
  */
 public record ReviewSummary(
-        Long reviewId,
+        int reviewId,
         String author,
         String subject
 ) {
